Fix write on closed channel error
In case of an error occuring in Fixup, there is a race condition where
we send an event after closing the event channel.

This is because the main task can complete (with an error) earlier than
other concurrent scheduled tasks running in the errgroup.

This adds a drain function to the errgroupScheduler to wait for all
goroutines to complete.

Signed-off-by: Simon Ferquel <simon.ferquel@docker.com>