Update docker/cli
-> 23446275646041f9b598d64c51be24d5d0e49376

This target commit is the one that copies `types.AuthConfig` from docker
api to docker cli.
By itself it's not an issue but it can be if `cnab-to-oci` is integrated
in a context where a newer version of `docker/cli` is required.
See https://github.com/docker/cnab-to-oci/pull/76#discussion_r335174252
for example, and the same kind of issue exists in `docker/app` where both
`cnab-to-oci` and a more recent `docker/cli` are used.

Signed-off-by: Yves Brissaud <yves.brissaud@docker.com>