allow to push local docker images by digest
In the case `contentDigest` is defined (for example after a build
by docker app) try to push local docker image from the digest.
In that case, the `image` of the service is not defined in the
bundle.json.

If `contentDigest` is empty, do the same thing as before, so try to resolve
and if it's not resolvable try to push.

The service name is now displayed in error messages to better understand
what's going on on failures.

--

A `internal.ImageClient` interface has been added to only deal with
methods from Docker's `ImageAPIClient` we need. Easier to mock for
example.

Signed-off-by: Yves Brissaud <yves.brissaud@docker.com>