Use relocation map during bundle fixup
I believe the relocation map was accidentally not used after
WithRelocationMap is set, and it news up an empty relocation map before
fixing the images in the bundle.

I have updated the fixup to use the relocation mapping set by the
caller, and fixed the mocks in the tests to account for the new code
flow.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>