Include digest when pulling a bundle
When pulling a bundle, return the digest of the bundle as well. This
allows clients to record the digest of a bundle that was pulled by a
tag, and use it for comparison later or allowing pulling from tag the
first time and then always pulling from the known digest on subsequent
bundle executions.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>