Skip duplicate image layers when copying
Some images have duplicate layers, for example docker/whalesay. This can be caused by DOCKERFILE commands that are executed but do not modify the filesystem. While it's not optimal for an image to do this, it shouldn't cause errors either.

When we copy the images used by a manifest into the bundle repository, we are creating a copy task for each layer resulting in multiple copy tasks attempting to copy the same layer. Only the first copy task works, and the others fail with "unexpected size" errors because the layer was already copied (fully or in part) in the previous task.

I've updated descriptorContentHandler to keep track of the layers, so that we can skip over duplicate layers.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>