Use same canonical json library as cnab-go
The cnab-go library that we use to work with bundles switched to a canonical json library with support for numeric types a while back and I forgot to update cnab-to-oci at the same time.

This updates cnab-to-oci to use the same library, github.com/cyberphone/json-canonicalization, and when possible just use the helper methods defined on cnab-go's Bundle type instead of doing the marshaling ourselves.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>